Extract `Carousel` to reusable component
