+++ src/components/books/RemoveBook.js
@@ -1,0 +1,3 @@
+const RemoveBook = () => (<div>Remove Book</div>);
+
+export default RemoveBook;
