+++ src/components/views/Carousel.js
@@ -1,15 +1,10 @@
-import { useSelector } from 'react-redux';
-import { Link } from 'react-router-dom';
-import { BsCaretRight, BsCaretLeft } from 'react-icons/bs';
+/* eslint-disable react/prop-types */
 import { useRef } from 'react';
-import BookView from './BookView';
-import styles from './Books.module.scss';
+import { BsCaretRight, BsCaretLeft } from 'react-icons/bs';
+import styles from './Carousel.module.scss';
 
-const Books = () => {
-  const { data: books, state } = useSelector((state) => state.booksReducer);
+const Carousel = ({ children }) => {
   const carouselEl = useRef(null);
-
-  if (state === 'loading') return <div>loading</div>;
 
   const handleClick = (dir) => {
     if (!carouselEl.current) return;
@@ -21,19 +16,7 @@ const Books = () => {
   return (
     <div className={styles.books}>
       <div ref={carouselEl} className={styles.transformed}>
-        { books.map(
-          ({
-            id, title, author, image,
-          }) => (
-            <Link to={`${id}`} key={title + author}>
-              <BookView
-                title={title}
-                author={author}
-                image={image}
-              />
-            </Link>
-          ),
-        )}
+        {children}
       </div>
       <button className={styles['nav-left']} aria-label="nav-left" type="button" onClick={() => handleClick('left')}><BsCaretLeft /></button>
       <button className={styles['nav-right']} aria-label="nav-right" type="button" onClick={() => handleClick('right')}><BsCaretRight /></button>
@@ -41,4 +24,4 @@ const Books = () => {
   );
 };
 
-export default Books;
+export default Carousel;
