+++ src/components/views/Carousel.module.scss
